Try to use pretty URLs not ending in /index.html
We'd like our internal URLs to point to locations like `../end/`, not
`../end/index.html`. But with Parcel.js this is actually surprisingly
hard to get working, as documented here:

https://github.com/parcel-bundler/parcel/issues/1186/

By adding the `parcel-resolve-ignore` module we can get things working,
since we index all of our inputs in the magic `_all.html` file. It adds
some additional output that's a little annoying but it's worth it, I
think.