First messy support for incremental builds
It is a total mess and I have broken actual output and it will take a
long time to un-break it, but the incremental-izing approach seems to
work! It has taken me so long to get here that I will commit and
checkpoint this even though there is a bunch of gross hackery here that
is temporary and will have to be reverted as we improve things.