Iterate the incremental build framework
There were various complex pieces that were written as specifically
referencing the pass1 operation, when really they could be made fully
generic with a bit of rearrangement. Also pass the cache as an argument
to various functions called throughout the TeX worker system, rather
than trying to save it as an Arc<Mutex<>>. It is very sensible for the
worker driver to have access to the cache so that it can know which
operations actually need to be performed.