.devcontainer: hack around Tectonic caching issues
When a build is executed with a pristine Tectonic local cache, the
different builds running in parallel run into problems with all trying
to update the cache simultaneously, leading to failures associated with
missing files.

The real solution is to make the Tectonic cache code more parallel-safe.
But as a quick workaround, do a single-threaded build on startup. This
will make sure that the cache is populated with everything we need
before starting the "watch". It is still possible for cache problems to
happen, so we still need to fix Tectonic. But this should get us going
in a kinda-reliable way.