src/tex_pass.rs: async-ify the subprocess comms
This gets hairy because we use a bunch of trait methods, where
async-in-trait only just became stable, and also the asset info that we
write to stdin in pass 2 is expected to be written synchronously by
spx2html, while we have an async child process. We can make it work, but
it involves buffering everything and is pretty gross. Hopefully we can
tidy this up over time.