Start working on bus-ifying individual file TeX outputs
We have to update the bus system to not be borrowing things, because the
futures will need to be able to be sent across threads if we're going to
use the async task pool to process all of the files. This is a little
annoying but ultimately makes sense -- borrows are a lot "riskier" in
async-land.