Update interaction with `yarn serve`
Renaming the build directory back and forth wasn't working reliably.
Which, to be honest, I probably should have expected. In a new approach,
the TeX build writes directly into `build/`, and then we duplicate its
files into a different directory, `serve/`, which is what `yarn serve`
actually monitors. This is not as "atomic" as I'd like, but at least we
can ensure that if the build fails, Yarn doesn't get sent a partial set
of files ... modulo I/O errors, which is an inesacapable possibility.