src/serve.rs: work to speed Parcel handoff by writing then renaming
Try to make the handoff to `parcel serve` even more direct by writing
the copies of the files that need updating into a temporary staging
area, then renaming them all in a group. That way we get all of the
"hard" I/O done before propagating any updates to the Parcel server.