Fix up Parcel's access to the search data JSON file
We were getting too-soon rebuilds in `serve` mode because
`web/SearchTool.vue` was referencing `../build/search_index.json.data`,
when in this scheme it should have been referencing the version in
`../serve/`. Of course, in build mode, the path shouldn't change.

We could fix this by tweaking the file access pattern of `serve` mode
(`build` -> `staging`, `serve` -> `build`), but I never really loved the
way the file in `web/` was referencing something in `build/` anyway.
Avoid that by propagating in the key information (the Parcel-munged
URL of the search index file) from the specific app build.

Unfortunately this turned out to be a lot harder than I'd have liked.
The final approach is fairly straightforward, but I had to try many
different methods to find one that would work with Parcel's bundling
scheme. At least the final approach is extensible if it turns out that
there are other such build-specific parameters that we need to propagate
into the code.

The end result of this is that if we edit a `.tex` source file, the
`yarn serve` process only does one rebuild, and it does it at the right
time. And so far the new copy-based approach seems like it is much more
reliable than the move-based one.