cls/pedia: fix TeX slugs with ligature-prone characters
I was totally confused because I was getting an unclear "undefined
control sequence" error with an entry whose name contained a hyphen.
Eventually I figured out that our \@noligs was causing problems because
it avoids ligatures by making the relevant characters, including `-`,
active. When we save the `-` in our token list and then expand it, we
get an error because the token is saved as having an active catcode, but
its definition doesn't exist.

This also helped me realize that I was seeing some typographic problems
because our mono font was set up to use TeX ligatures. So a character
like the backtick would be converted to a left single curly quote even
though if it's being rendered in a mono font it should almost definitely
be an actual backtick. So, stop doing that. Then we can activate the
mono font during our verbatim processing, which should prevent
ligatures from getting activated in that context.

It might, in fact, make sense to define "different" fonts that use the
same typefaces but have different ligature settings, and we can do this.
But to first order, it seems like the sans/serif fonts should have
regular TeX ligatures activated and the mono font shouldn't.